text: move StyleRange to new "style" module
